requirements.txt
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==6.5.0
idna==3.4
iniconfig==1.1.1
packaging==22.0
pluggy==1.0.0
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
requests==2.28.1
urllib3==1.26.13
langgraph-cli[inmem]
duckduckgo-search
ddgs
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain
langchain-ollama
notebook
wikipedia
python-dotenv
chromadb
pandas
ipykernel
iprogress
ipywidgets
jupyter
jupyterlab